Rename s/_/-/g src/**/*.js test/**/*.js
- Per @scottgonzalez review;
