+++ package.json
@@ -49,7 +49,7 @@
     "Tobie Langel <tobie.langel@gmail.com> (http://tobielangel.com)",
     "Boris Moore <borismoore@gmail.com> (http://www.borismoore.com)"
   ],
-  "main": "./dist/globalize_allinone_node.js",
+  "main": "./dist/globalize-allinone-node.js",
   "repository": {
     "type": "git",
     "url": "git://github.com/jquery/globalize.git"
