+++ src/common/get-locale.js
