+++ src/date/expand-pattern.js
