+++ src/date/pattern-re.js
