+++ src/util/array/is-array.js
